YEL-76 [refactor] service refactoring
